pyproject.toml
[project]
name = "multi-modal-search-engine"
version = "0.0.1"
description = ""
authors = [
    {name = "Suha Mokalla",email = "suhareddy.mokalla@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "streamlit (>=1.43.2,<2.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "polars (>=1.24.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
